% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColorBar.R
\name{addColorBar}
\alias{addColorBar}
\title{Add a color bar (useful for heatmaps or maps)}
\usage{
addColorBar(x, y, z, color = rainbow(100))
}
\arguments{
\item{x}{x-position where to plot the color ramp}

\item{y}{y-position where to plot the color ramp}

\item{z}{vector of values for the color ramp (doesn't have to be evenly space sequence)}
}
\value{
nothing
}
\description{
Add a color bar (useful for heatmaps or maps)
}
