% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArea.R
\name{getArea}
\alias{getArea}
\title{Calculate the area formed by a bunch of points}
\usage{
getArea(x, y, convexHull = TRUE)
}
\arguments{
\item{x}{vector of x values}

\item{y}{vector of y values}

\item{convexHull}{(boolean) should the area be a convex hull?}
}
\description{
Calculate the area formed by a bunch of points
}
