% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMidpoint.R
\name{getMidpoint}
\alias{getMidpoint}
\title{Get midpoint of factor level (useful for plotting)}
\usage{
getMidpoint(x, breaks = NULL)
}
\arguments{
\item{x}{vector of values to cut}

\item{breaks}{number of breaks}
}
\value{
midpoint for each point in x
}
\description{
Get midpoint of factor level (useful for plotting)
}
