% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRange.R
\name{getRange}
\alias{getRange}
\title{Get range for factor level (useful for plotting)}
\usage{
getRange(x, breaks = NULL)
}
\arguments{
\item{x}{vector of values to cut}

\item{breaks}{number of breaks}
}
\value{
range of values in that factor level for each value in x
}
\description{
Get range for factor level (useful for plotting)
}
