% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpectral.R
\name{getSpectral}
\alias{getSpectral}
\title{Calculate the spectral radius of a network (in interaction matrix format)}
\usage{
getSpectral(x)
}
\arguments{
\item{x}{interaction matrix}
}
\description{
Calculate the spectral radius of a network (in interaction matrix format)
}
