% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodColors.R
\name{goodColors}
\alias{goodColors}
\title{Import some color palettes}
\usage{
goodColors(index = 1, alpha = 0.5)
}
\arguments{
\item{index}{which palette do you want? (1-7)}

\item{alpha}{transparency channel}
}
\value{
color palette
}
\description{
Import some color palettes
}
