% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorPlot.R
\name{makeCorPlot}
\alias{makeCorPlot}
\title{Create correlation matrix plot with Spearman's rank corr output}
\usage{
makeCorPlot(x, colorz = "dodgerblue", varNames = NULL)
}
\arguments{
\item{x}{a matrix or data.frame containing variables of interest}

\item{colorz}{a single color to use for gradient shading}

\item{varNames}{names to use as labels (default is to use colnames of x)}
}
\value{
nothing but a plot
}
\description{
Create correlation matrix plot with Spearman's rank corr output
}
