% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummy.R
\name{makeDummy}
\alias{makeDummy}
\title{Make a dummy variable from a categorical or factor variable}
\usage{
makeDummy(x)
}
\arguments{
\item{x}{a vector of type character or factor}
}
\value{
a binary matrix containing dummy variables
}
\description{
Make a dummy variable from a categorical or factor variable
}
