% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot a graph object}
\usage{
plotGraph(
  g,
  lay = layout_nicely(g),
  colz = 1:5,
  nodeSize = 10,
  colVec = NULL,
  edgeWidth = 1
)
}
\arguments{
\item{g}{graph object}

\item{colz}{colors to use}

\item{nodeSize}{Size of nodes}

\item{colVec}{What vertex attribute should color nodes? (default = NULL)}

\item{edgeWidth}{edge weights e.g., E(g)$weight? (default=1)}
}
\value{
a graph plot
}
\description{
Plot a graph object
}
