% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presFig.R
\name{presFig}
\alias{presFig}
\title{Make the default color of axes and labels white (for presentations with dark background)}
\usage{
presFig(...)
}
\arguments{
\item{...}{additional par arguments (if any)}
}
\value{
nothing
}
\description{
Make the default color of axes and labels white (for presentations with dark background)
}
