% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegments.R
\name{plotSegments}
\alias{plotSegments}
\title{Plot a linear fit within the bounds of the data, and insert fit stats into plot}
\usage{
plotSegments(
  y,
  x,
  forceIntercept = NULL,
  legPos = NULL,
  justMod = FALSE,
  textCol = 1,
  ...
)
}
\arguments{
\item{y}{a vector}

\item{x}{a vector}

\item{forceIntercept}{(default = NULL) what is the intercept value to use (useful 
for relationships that need to be forced through origin}

\item{legPos}{position of the legend (which is the slope and p-value; default=NULL)
other values are "topleft", "bottomleft", "topright", "bottomright".}

\item{justMod}{if TRUE, no segment is drawn, but the model object is returned}

\item{textCol}{color of line and text.}

\item{...}{other graphical parameters taht will be handed to the lines function}
}
\value{
nothing
}
\description{
Plot a linear fit within the bounds of the data, and insert fit stats into plot
}
